+++ packages/at_server_spec/lib/src/connection/inbound_connection.dart
@@ -1,7 +1,8 @@
 import 'dart:io';
+import 'package:at_server_spec/src/at_rate_limiter/at_rate_limiter.dart';
 import 'package:at_server_spec/src/connection/at_connection.dart';
 
-abstract class InboundConnection extends AtConnection {
+abstract class InboundConnection extends AtConnection implements AtRateLimiter {
   ///Returns true if remote socket and remote port of this and connection matches
   bool equals(InboundConnection connection);